--- a/Cargo.toml
+++ b/Cargo.toml
@@ -24,7 +24,6 @@ rand_core = "0.5"
 byteorder = "1"
 log = "0.4.8"
 lazy_static = "1.4.0"
-ocl = { version = "0.19.4", package = "fil-ocl", optional = true }
 ff-cl-gen = { version = "0.1.1", optional = true }
 fs2 = { version = "0.4.3", optional = true }
 rand = "0.7"
@@ -32,6 +31,7 @@ rayon = "1.3.0"
 memmap = "0.7.0"
 thiserror = "1.0.10"
 ahash = "0.3.4"
+rust-gpu-tools = { version = "0.1.0", optional = true }
 
 [dev-dependencies]
 hex-literal = "0.2"
@@ -42,7 +42,7 @@ criterion = "0.3.2"
 
 [features]
 default = ["groth16", "multicore"]
-gpu = ["ocl", "ff-cl-gen", "fs2"]
+gpu = ["rust-gpu-tools", "ff-cl-gen", "fs2"]
 groth16 = ["paired"]
 multicore = ["futures-cpupool", "crossbeam", "num_cpus"]