--- a/src/domain.rs
+++ b/src/domain.rs
@@ -559,11 +559,11 @@ fn parallel_fft_consistency() {
     test_consistency::<Bls12, _>(rng);
 }
 
-pub fn create_fft_kernel<E>(log_d: usize, priority: bool) -> Option<gpu::FFTKernel<E>>
+pub fn create_fft_kernel<E>(_log_d: usize, priority: bool) -> Option<gpu::FFTKernel<E>>
 where
     E: Engine,
 {
-    match gpu::FFTKernel::create(1 << log_d, priority) {
+    match gpu::FFTKernel::create(priority) {
         Ok(k) => {
             info!("GPU FFT kernel instantiated!");
             Some(k)
@@ -594,7 +594,7 @@ mod tests {
 
         let worker = Worker::new();
         let log_cpus = worker.log_num_cpus();
-        let mut kern = gpu::FFTKernel::create(1 << 24, false).expect("Cannot initialize kernel!");
+        let mut kern = gpu::FFTKernel::create(false).expect("Cannot initialize kernel!");
 
         for log_d in 1..25 {
             let d = 1 << log_d;