--- a/src/gpu/fft.rs
+++ b/src/gpu/fft.rs
@@ -1,29 +1,24 @@
-use crate::gpu::get_platform;
 use crate::gpu::{
     error::{GPUError, GPUResult},
-    get_devices, locks, sources, structs,
+    locks, sources,
 };
 use ff::Field;
 use log::info;
-use ocl::{Buffer, MemFlags, ProQue};
 use paired::Engine;
+use rust_gpu_tools::*;
 use std::cmp;
 
-// NOTE: Please read `structs.rs` for an explanation for unsafe transmutes of this code!
-
 const LOG2_MAX_ELEMENTS: usize = 32; // At most 2^32 elements is supported.
-const MAX_RADIX_DEGREE: u32 = 8; // Radix256
-const MAX_LOCAL_WORK_SIZE_DEGREE: u32 = 7; // 128
+const MAX_LOG2_RADIX: u32 = 8; // Radix256
+const MAX_LOG2_LOCAL_WORK_SIZE: u32 = 7; // 128
 
 pub struct FFTKernel<E>
 where
     E: Engine,
 {
-    proque: ProQue,
-    fft_src_buffer: Buffer<structs::PrimeFieldStruct<E::Fr>>,
-    fft_dst_buffer: Buffer<structs::PrimeFieldStruct<E::Fr>>,
-    fft_pq_buffer: Buffer<structs::PrimeFieldStruct<E::Fr>>,
-    fft_omg_buffer: Buffer<structs::PrimeFieldStruct<E::Fr>>,
+    program: opencl::Program,
+    pq_buffer: opencl::Buffer<E::Fr>,
+    omegas_buffer: opencl::Buffer<E::Fr>,
     _lock: locks::GPULock, // RFC 1857: struct fields are dropped in the same order as they are declared.
     priority: bool,
 }
@@ -32,172 +27,123 @@ impl<E> FFTKernel<E>
 where
     E: Engine,
 {
-    pub fn create(n: u32, priority: bool) -> GPUResult<FFTKernel<E>> {
+    pub fn create(priority: bool) -> GPUResult<FFTKernel<E>> {
         let lock = locks::GPULock::lock();
         let src = sources::kernel::<E>();
 
-        let platform = get_platform(None)?;
-        info!("Platform selected: {}", platform.name()?);
-
-        let devices = get_devices(&platform).unwrap_or_default();
+        let devices = opencl::Device::all()?;
         if devices.is_empty() {
             return Err(GPUError::Simple("No working GPUs found!"));
         }
 
         // Select the first device for FFT
-        let device = devices[0];
+        let device = devices[0].clone();
 
-        let pq = ProQue::builder()
-            .platform(platform)
-            .device(device)
-            .src(src)
-            .dims(n)
-            .build()?;
-
-        let srcbuff = Buffer::builder()
-            .queue(pq.queue().clone())
-            .flags(MemFlags::new().read_write())
-            .len(n)
-            .build()?;
-        let dstbuff = Buffer::builder()
-            .queue(pq.queue().clone())
-            .flags(MemFlags::new().read_write())
-            .len(n)
-            .build()?;
-        let pqbuff = Buffer::builder()
-            .queue(pq.queue().clone())
-            .flags(MemFlags::new().read_write())
-            .len(1 << MAX_RADIX_DEGREE >> 1)
-            .build()?;
-        let omgbuff = Buffer::builder()
-            .queue(pq.queue().clone())
-            .flags(MemFlags::new().read_write())
-            .len(LOG2_MAX_ELEMENTS)
-            .build()?;
+        let program = opencl::Program::from_opencl(device, &src)?;
+        let pq_buffer = program.create_buffer::<E::Fr>(1 << MAX_LOG2_RADIX >> 1)?;
+        let omegas_buffer = program.create_buffer::<E::Fr>(LOG2_MAX_ELEMENTS)?;
 
         info!("FFT: 1 working device(s) selected.");
-        info!("FFT: Device 0: {}", pq.device().name()?);
+        info!("FFT: Device 0: {}", program.device().name());
 
         Ok(FFTKernel {
-            proque: pq,
-            fft_src_buffer: srcbuff,
-            fft_dst_buffer: dstbuff,
-            fft_pq_buffer: pqbuff,
-            fft_omg_buffer: omgbuff,
+            program,
+            pq_buffer,
+            omegas_buffer,
             _lock: lock,
             priority,
         })
     }
 
     /// Peforms a FFT round
-    /// * `lgn` - Specifies log2 of number of elements
-    /// * `lgp` - Specifies log2 of `p`, (http://www.bealto.com/gpu-fft_group-1.html)
+    /// * `log_n` - Specifies log2 of number of elements
+    /// * `log_p` - Specifies log2 of `p`, (http://www.bealto.com/gpu-fft_group-1.html)
     /// * `deg` - 1=>radix2, 2=>radix4, 3=>radix8, ...
     /// * `max_deg` - The precalculated values pq` and `omegas` are valid for radix degrees up to `max_deg`
     fn radix_fft_round(
         &mut self,
-        lgn: u32,
-        lgp: u32,
+        src_buffer: &opencl::Buffer<E::Fr>,
+        dst_buffer: &opencl::Buffer<E::Fr>,
+        log_n: u32,
+        log_p: u32,
         deg: u32,
         max_deg: u32,
-        in_src: bool,
     ) -> GPUResult<()> {
         if locks::PriorityLock::should_break(self.priority) {
             return Err(GPUError::GPUTaken);
         }
 
-        let n = 1u32 << lgn;
-        let lwsd = cmp::min(deg - 1, MAX_LOCAL_WORK_SIZE_DEGREE);
-        let kernel = self
-            .proque
-            .kernel_builder("radix_fft")
-            .global_work_size([n >> deg << lwsd])
-            .local_work_size(1 << lwsd)
-            .arg(if in_src {
-                &self.fft_src_buffer
-            } else {
-                &self.fft_dst_buffer
-            })
-            .arg(if in_src {
-                &self.fft_dst_buffer
-            } else {
-                &self.fft_src_buffer
-            })
-            .arg(&self.fft_pq_buffer)
-            .arg(&self.fft_omg_buffer)
-            .arg_local::<structs::PrimeFieldStruct<E::Fr>>(1 << deg)
-            .arg(n)
-            .arg(lgp)
-            .arg(deg)
-            .arg(max_deg)
-            .build()?;
-        unsafe {
-            kernel.enq()?;
-        } // Running a GPU kernel is unsafe!
+        let n = 1u32 << log_n;
+        let local_work_size = 1 << cmp::min(deg - 1, MAX_LOG2_LOCAL_WORK_SIZE);
+        let global_work_size = (n >> deg) * local_work_size;
+        let kernel = self.program.create_kernel(
+            "radix_fft",
+            global_work_size as usize,
+            Some(local_work_size as usize),
+        );
+        call_kernel!(
+            kernel,
+            src_buffer,
+            dst_buffer,
+            &self.pq_buffer,
+            &self.omegas_buffer,
+            opencl::LocalBuffer::<E::Fr>::new(1 << deg),
+            n,
+            log_p,
+            deg,
+            max_deg
+        )?;
         Ok(())
     }
 
     /// Share some precalculated values between threads to boost the performance
-    fn setup_pq(&mut self, omega: &E::Fr, n: usize, max_deg: u32) -> ocl::Result<()> {
+    fn setup_pq_omegas(&mut self, omega: &E::Fr, n: usize, max_deg: u32) -> GPUResult<()> {
         // Precalculate:
         // [omega^(0/(2^(deg-1))), omega^(1/(2^(deg-1))), ..., omega^((2^(deg-1)-1)/(2^(deg-1)))]
-        let mut tpq = vec![structs::PrimeFieldStruct::<E::Fr>::default(); 1 << max_deg >> 1];
-        let pq = unsafe {
-            std::mem::transmute::<&mut [structs::PrimeFieldStruct<E::Fr>], &mut [E::Fr]>(&mut tpq)
-        };
-        let tw = omega.pow([(n >> max_deg) as u64]);
+        let mut pq = vec![E::Fr::zero(); 1 << max_deg >> 1];
+        let twiddle = omega.pow([(n >> max_deg) as u64]);
         pq[0] = E::Fr::one();
         if max_deg > 1 {
-            pq[1] = tw;
+            pq[1] = twiddle;
             for i in 2..(1 << max_deg >> 1) {
                 pq[i] = pq[i - 1];
-                pq[i].mul_assign(&tw);
+                pq[i].mul_assign(&twiddle);
             }
         }
-        self.fft_pq_buffer.write(&tpq).enq()?;
+        self.pq_buffer.write_from(&pq)?;
 
         // Precalculate [omega, omega^2, omega^4, omega^8, ..., omega^(2^31)]
-        let mut tom = vec![structs::PrimeFieldStruct::<E::Fr>::default(); 32];
-        let om = unsafe {
-            std::mem::transmute::<&mut [structs::PrimeFieldStruct<E::Fr>], &mut [E::Fr]>(&mut tom)
-        };
-        om[0] = *omega;
+        let mut omegas = vec![E::Fr::zero(); 32];
+        omegas[0] = *omega;
         for i in 1..LOG2_MAX_ELEMENTS {
-            om[i] = om[i - 1].pow([2u64]);
+            omegas[i] = omegas[i - 1].pow([2u64]);
         }
-        self.fft_omg_buffer.write(&tom).enq()?;
+        self.omegas_buffer.write_from(&omegas)?;
 
         Ok(())
     }
 
     /// Performs FFT on `a`
     /// * `omega` - Special value `omega` is used for FFT over finite-fields
-    /// * `lgn` - Specifies log2 of number of elements
-    pub fn radix_fft(&mut self, a: &mut [E::Fr], omega: &E::Fr, lgn: u32) -> GPUResult<()> {
-        let n = 1 << lgn;
+    /// * `log_n` - Specifies log2 of number of elements
+    pub fn radix_fft(&mut self, a: &mut [E::Fr], omega: &E::Fr, log_n: u32) -> GPUResult<()> {
+        let n = 1 << log_n;
+        let mut src_buffer = self.program.create_buffer::<E::Fr>(n)?;
+        let mut dst_buffer = self.program.create_buffer::<E::Fr>(n)?;
 
-        let ta = unsafe {
-            std::mem::transmute::<&mut [E::Fr], &mut [structs::PrimeFieldStruct<E::Fr>]>(a)
-        };
+        let max_deg = cmp::min(MAX_LOG2_RADIX, log_n);
+        self.setup_pq_omegas(omega, n, max_deg)?;
 
-        let max_deg = cmp::min(MAX_RADIX_DEGREE, lgn);
-        self.setup_pq(omega, n, max_deg)?;
-
-        self.fft_src_buffer.write(&*ta).enq()?;
-        let mut in_src = true;
-        let mut lgp = 0u32;
-        while lgp < lgn {
-            let deg = cmp::min(max_deg, lgn - lgp);
-            self.radix_fft_round(lgn, lgp, deg, max_deg, in_src)?;
-            lgp += deg;
-            in_src = !in_src; // Destination of this FFT round is source of the next round.
+        src_buffer.write_from(&*a)?;
+        let mut log_p = 0u32;
+        while log_p < log_n {
+            let deg = cmp::min(max_deg, log_n - log_p);
+            self.radix_fft_round(&src_buffer, &dst_buffer, log_n, log_p, deg, max_deg)?;
+            log_p += deg;
+            std::mem::swap(&mut src_buffer, &mut dst_buffer);
         }
-        if in_src {
-            self.fft_src_buffer.read(ta).enq()?;
-        } else {
-            self.fft_dst_buffer.read(ta).enq()?;
-        }
-        self.proque.finish()?; // Wait for all commands in the queue (Including read command)
+
+        src_buffer.read_into(a)?;
 
         Ok(())
     }