--- a/src/gpu/mod.rs
+++ b/src/gpu/mod.rs
@@ -21,12 +21,6 @@ mod utils;
 pub use self::utils::*;
 
 #[cfg(feature = "gpu")]
-mod structs;
-
-#[cfg(feature = "gpu")]
-pub use self::structs::*;
-
-#[cfg(feature = "gpu")]
 mod fft;
 
 #[cfg(feature = "gpu")]