Use `rust-gpu-tools` instead of lower-level `ocl` library.

**Notes:**
 - First run: No performance regression seen in 512MB/32GB Window-PoSt circuits
 - Second run: ~1% (5:03 -> 5:06s) performance regression in 32GB Window-PoSt
 - ~1.8% speedup seen in sealing part (39:57 -> 39:14) (Not totally reliable, since we run it only a single time)

**Pros:**
 - Brings a much cleaner API for developing OpenCL applications
 - Unnecessary transmutes are not needed anymore
 - Automatic program caching for AMD cards
 - Buffers are now allocated only when they are needed (This could make the integration of GPU scheduler easier, in the future)

**Cons:**
 - ~1% performance regression seen on dummy exponentiation circuit with 33mm elements
 - ~1% performance regression 32GB Window-PoSt (5:03 -> 5:06s)