@keyvank 2020-07-26
## Correctness check
The OpenCL codes and host logic has not been changed. Correctness check tests are still passing on circuits with millions of constraints.
The underlying `ocl` buffer of `rust-gpu-tools` is `Buffer<u8>` instead of `Buffer<E::Fr>` or `Buffer<E::G1/2>`, I was slightly worried that using a lower-level buffer and having our own implementation of generic buffers instead of using the generic buffers provided by `ocl` library, could lead to padding/alignment issues and wrong results, but after testing it a few times on different machines/platforms I'm almost confident that it is working correctly.

## Performance reports
- Performance difference of multiexps of 2^27 (~134mm) elements
    * First run: Previously: 45560ms This PR: 46444ms
    * Second run: Previously: 45974ms This PR: 46608ms
    * Third run: Previously: 45744ms This PR: 46573ms
    * Average: Previously: 45759ms This PR: 46541ms
    * Regression: **_~1.7% slower!_** 
- Performance difference of complete SNARK proofs for 33mm (2^25) constraint circuits (Average 4 runs)
    * Previously: 112.5s This PR: 113.7s
    * Regression: _**~1.1% slower!**_
- No considerable performance regression detected on Window-PoST proving of 512MB sectors **_(I estimate that we will have a 1-1.5% regression for bigger sector sizes/circuits)_**