@keyvank 2020-08-11
Another run (Window PoSt 32GB + Sealing):

## master branch:
First bellman run (Sealing?): 39:57s
Second bellman run: 5:03s

## This branch
First bellman run (Sealing?): 39:14s (~1.8% faster)
Second bellman run: 5:06s (~1% slower)